mysite/requirements.txt
arrow==1.2.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
Django==4.1.4
django-browser-reload==1.6.0
django-tailwind==3.4.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
Pillow==9.3.0
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==7.0.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
urllib3==1.26.13
